move logs to _logs (#2240)
* move logs to _logs

* fix lint
