+++ docs/sdk/logs.severity.rst
@@ -1,7 +1,7 @@
-opentelemetry.sdk.logs.severity
-===============================
+opentelemetry.sdk._logs.severity
+================================
 
-.. automodule:: opentelemetry.sdk.logs.severity
+.. automodule:: opentelemetry.sdk._logs.severity
     :members:
     :undoc-members:
     :show-inheritance:
