+++ exporter/opentelemetry-exporter-otlp-proto-grpc/tests/logs/test_otlp_logs_exporter.py
@@ -42,9 +42,11 @@ from opentelemetry.proto.logs.v1.logs_pb2 import ResourceLogs
 from opentelemetry.proto.resource.v1.resource_pb2 import (
     Resource as OTLPResource,
 )
-from opentelemetry.sdk.logs import LogData, LogRecord
-from opentelemetry.sdk.logs.export import LogExportResult
-from opentelemetry.sdk.logs.severity import SeverityNumber as SDKSeverityNumber
+from opentelemetry.sdk._logs import LogData, LogRecord
+from opentelemetry.sdk._logs.export import LogExportResult
+from opentelemetry.sdk._logs.severity import (
+    SeverityNumber as SDKSeverityNumber,
+)
 from opentelemetry.sdk.resources import Resource as SDKResource
 from opentelemetry.sdk.util.instrumentation import InstrumentationInfo
 from opentelemetry.trace import TraceFlags
