+++ opentelemetry-sdk/setup.cfg
@@ -55,7 +55,7 @@ opentelemetry_tracer_provider =
 opentelemetry_traces_exporter =
     console = opentelemetry.sdk.trace.export:ConsoleSpanExporter
 opentelemetry_log_emitter_provider =
-    sdk_log_emitter_provider = opentelemetry.sdk.logs:LogEmitterProvider
+    sdk_log_emitter_provider = opentelemetry.sdk._logs:LogEmitterProvider
 opentelemetry_id_generator =
     random = opentelemetry.sdk.trace.id_generator:RandomIdGenerator
 opentelemetry_environment_variables =
