+++ opentelemetry-sdk/src/opentelemetry/sdk/_logs/severity.py
